Fix logging

+++ etc/init/osx/0_awscli.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
-curl "https://awscli.amazonaws.com/AWSCLIV2.pkg" -o "AWSCLIV2.pkg"
-sudo installer -pkg AWSCLIV2.pkg -target /
+if ! is_exists "aws"; then
+  curl "https://awscli.amazonaws.com/AWSCLIV2.pkg" -o "AWSCLIV2.pkg"
+  sudo installer -pkg AWSCLIV2.pkg -target /
+  log_pass "Success to initialize AWS CLI"
+else
+  log_warn "AWS CLI is already installed."
+fi
